+++ CHANGELOG.md
@@ -9,6 +9,8 @@ Yii Framework 2 redis extension Change Log
 - Enh #287: Applying Yii2 coding standards (@s1lver)
 - Fix #281: PHP 7.4 compatibility is now fixed (@s1lver)
 - Fix #291: Prevent infinite loop in `parseResponse()` when `fread()` returns empty string on broken socket (@dkostik)
+- Fix #292: Fixed `fwrite()` supplied resource is not a valid stream (@s1lver)
+- Enh #292: Bump version `predis/predis` to `3.4` (@s1lver)
 
 2.1.0 December 25, 2025
 -----------------------